requirements.txt
# ============================================
# SAP Transformation Management Platform
# requirements.txt
# Python 3.11+
# ============================================
# Sprint 1 – Foundation dependencies
# ============================================

# --- Web Framework ---
Flask==3.1.0

# --- ORM & Database ---
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg[binary]==3.2.4

# --- Utilities ---
Flask-CORS==5.0.0
python-dotenv==1.0.1
Flask-Limiter==3.8.0
openpyxl==3.1.5
weasyprint==62.3

# --- Cache / Queue ---
redis==5.2.1

# --- Production Server ---
gunicorn==23.0.0

# --- Auth & Security ---
PyJWT==2.9.0
bcrypt==4.2.1
email-validator==2.2.0

# --- Testing ---
pytest==8.3.4
pytest-flask==1.3.0

# --- Linting (CI) ---
ruff>=0.8.0

# ============================================
# Sprint 7 – AI Infrastructure (optional)
# ============================================
PyYAML==6.0.2

# ============================================
# Sprint 7 – SSO / OIDC / SAML
# ============================================
Authlib>=1.3.0
httpx>=0.27.0

# ============================================
# Gemini Free-Tier Demo (google-genai SDK)
# ============================================
google-genai>=1.0.0
psycopg2-binary==2.9.11
